Add ParameterManager and required dependencies
This class was created with the intention that it be used to easily
load a XML file and query it for entries based on a known node
heirarchy. Since the parameters used for this project use
well-understood patterns, this strategy is useful for navigating
a set of parameters in an explicit, user-friendly manner.

A ParameterManager object is created by a controlling entity
(probably within an executable's main() function or some other
high-level entity). The parameter manager object contains a ptr/ref
to the XML tree, and can traverse the tree according to an XPath
provided by the class' client.

For classes which require an additional parameter file to be read
(as is the case for the existing FixedLayout class), a new,
stack-scoped or heap-allocated instance of the class can be created
to parse that file as needed.

Language spec: https://www.w3.org/TR/1999/REC-xpath-19991116/
XPath tutorial: https://www.w3schools.com/xml/xpath_syntax.asp

For a syntax and usage example, see
https://github.com/etcadinfinitum/BrainGrid/blob/parameter-refactor/Tests/,
particularly the tests_ParameterManager.cpp file.